Add some intelligent paths resolution Add test schema Tests prep

+++ src/Queil.Gmox.DotNet.Cli/Program.fs
@@ -4,7 +4,6 @@ open Queil.Gmox.App
 open Queil.Gmox.CliArgs
 open Queil.Gmox.Core
 open Queil.Gmox.Infra
-open Saturn
 open System.IO
 open System.Reflection
 open Fake.Core
@@ -13,6 +12,7 @@ type ProjInfo = {
   Name: string
   Path: string
   CompileOutputPath: string
+  AllProtoFiles: string list
   Options: Options
 }
 with member x.OutputAssemblyFullPath () = Path.Combine(x.Path, x.CompileOutputPath, $"{x.Name}.dll")
@@ -20,15 +20,14 @@ with member x.OutputAssemblyFullPath () = Path.Combine(x.Path, x.CompileOutputPa
 let createTempProj (info: ProjInfo) =
 
   let csProjPath = Path.Combine(info.Path, $"{info.Name}.csproj")
+  let protosRoot = 
+    match info.Options.ProtoRoot with
+    | Some path -> $"ProtoRoot=\"{path}\""
+    | None -> ""
 
   let importCompiles indent =
-    info.Options.ImportPaths
-    |> Seq.map (fun p -> $"""{indent}<Protobuf Include="{p}" GrpcServices="None" ProtoRoot="{info.Options.ProtoRoot}" />""")
-    |> String.concat System.Environment.NewLine
-
-  let protos indent =
-    info.Options.Proto
-    |> Seq.map (fun p -> $"""{indent}<Protobuf Include="{p}" GrpcServices="Server" ProtoRoot="{info.Options.ProtoRoot}" />""")
+    info.AllProtoFiles
+    |> Seq.map (fun p -> $"""{indent}<Protobuf Include="{p}" GrpcServices="Server" {protosRoot} />""")
     |> String.concat System.Environment.NewLine
 
   let template = $"""<Project Sdk="Microsoft.NET.Sdk">
@@ -47,7 +46,6 @@ let createTempProj (info: ProjInfo) =
 
 <ItemGroup>
 {"  " |> importCompiles}
-{"  " |> protos}
 </ItemGroup>
 
 </Project>"""
@@ -65,34 +63,55 @@ let tempProjPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName())
 let dir = Directory.CreateDirectory(tempProjPath)
 
 try
-  try
-    let opts =
-      System.Environment.GetCommandLineArgs().[1..]
-      |> parseOptions
+  let opts =
+    System.Environment.GetCommandLineArgs().[1..]
+    |> parseOptions
+    |> fun opts ->
+      let asmLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)
+      let googleIncludesPath = Path.Combine(asmLocation, "include")
+      {
+        opts with
+          ImportPaths = [
+            yield! opts.ImportPaths
+            yield googleIncludesPath
+            yield! opts.Proto |> List.map (Path.GetDirectoryName)
+            match opts.ProtoRoot with
+            | Some path -> yield path
+            | None -> ()
+          ]
+      }
 
-    printfn "%A" opts
+  let allFiles = 
+    opts.Proto
+    |> Seq.collect (fun proto -> Froto.Parser.Parse.loadFromFile proto opts.ImportPaths)
+    |> Seq.map fst
+    |> Seq.distinct
+    |> Seq.filter (fun p -> not <| p.Contains("/include/google/protobuf/"))
+    |> Seq.toList
 
-    let projInfo = {
-      Name = "grpc"
-      Path = tempProjPath
-      CompileOutputPath = "out"
-      Options = opts
-    }
+  let projInfo = {
+    Name = "proto-gen"
+    Path = tempProjPath
+    CompileOutputPath = "out"
+    AllProtoFiles = allFiles
+    Options = opts
+  }
 
-    projInfo |> createTempProj
-    |> Proc.run
-    |> ignore
+  if opts.DebugMode then printfn "%A" projInfo
 
-    let asm = Assembly.LoadFile(projInfo.OutputAssemblyFullPath ())
+  projInfo |> createTempProj
+  |> Proc.run
+  |> ignore
 
-    runGmox (app {
-      Services = Grpc.servicesFromAssembly asm |> Seq.map Emit.makeImpl |> Seq.toList
-      StubPreloadDir = opts.StubsDir
-    })
-  with
-    | :? Argu.ArguParseException as p ->
-      printfn "%s" p.Message
-    | e -> printfn "%s" (e.ToString())
-finally
-  dir.Delete(true)
-  ()
+  let asm = Assembly.LoadFile(projInfo.OutputAssemblyFullPath ())
+
+  runGmox (app {
+    Services = Grpc.servicesFromAssembly asm |> Seq.map Emit.makeImpl |> Seq.toList
+    StubPreloadDir = opts.StubsDir
+  })
+
+  if not <| opts.DebugMode then dir.Delete(true)
+with
+  | :? Argu.ArguParseException as p ->
+    printfn "%s" p.Message
+  | e -> printfn "%s" (e.ToString())
